Move 'stars' feature implementation to Base2DScreen.
